changes EmptyCommand to UnknownCommand
