+++ server/src/main/java/com/battleships/Commands/CommandsImpl/EndGame.java
@@ -1,18 +1,19 @@
-package com.battleships.Commands;
+package com.battleships.Commands.CommandsImpl;
 
 import com.battleships.commands.AbstractCommand;
 import com.battleships.commands.CommandType;
 import com.battleships.commands.Message;
 import com.battleships.player.ConnectedPlayers;
 
-public class SetupCompletedCommand extends AbstractCommand {
-    public SetupCompletedCommand(Object value) {
+public class EndGame extends AbstractCommand {
+    public EndGame(Object value) {
         super(value);
     }
 
     @Override
     public void execute(ConnectedPlayers connectedPlayers) {
-        connectedPlayers.sendToInactive(new Message(CommandType.SETUP_COMPLETED, true));
+        Message message = new Message(CommandType.END_GAME, value);
+        connectedPlayers.sendToInactive(message);
         connectedPlayers.switchActive();
     }
 }
