+++ server/src/main/java/com/battleships/Commands/CommandsImpl/PlayerHitResult.java
@@ -1,19 +1,19 @@
-package com.battleships.Commands;
+package com.battleships.Commands.CommandsImpl;
 
 import com.battleships.commands.AbstractCommand;
 import com.battleships.commands.CommandType;
 import com.battleships.commands.Message;
 import com.battleships.player.ConnectedPlayers;
 
-public class EndGame extends AbstractCommand {
-    public EndGame(Object value) {
+public class PlayerHitResult extends AbstractCommand {
+    public PlayerHitResult(Object value) {
         super(value);
     }
 
     @Override
     public void execute(ConnectedPlayers connectedPlayers) {
-        Message message = new Message(CommandType.END_GAME, value);
-        connectedPlayers.sendToInactive(message);
+        Message shotMessage = new Message(CommandType.HIT, value);
+        connectedPlayers.sendToInactive(shotMessage);
         connectedPlayers.switchActive();
     }
 }
