+++ server/src/main/java/com/battleships/Commands/CommandsImpl/SetupCompletedCommand.java
@@ -1,19 +1,18 @@
-package com.battleships.Commands;
+package com.battleships.Commands.CommandsImpl;
 
 import com.battleships.commands.AbstractCommand;
 import com.battleships.commands.CommandType;
 import com.battleships.commands.Message;
 import com.battleships.player.ConnectedPlayers;
 
-public class PlayerHitResult extends AbstractCommand {
-    public PlayerHitResult(Object value) {
+public class SetupCompletedCommand extends AbstractCommand {
+    public SetupCompletedCommand(Object value) {
         super(value);
     }
 
     @Override
     public void execute(ConnectedPlayers connectedPlayers) {
-        Message shotMessage = new Message(CommandType.HIT, value);
-        connectedPlayers.sendToInactive(shotMessage);
+        connectedPlayers.sendToInactive(new Message(CommandType.SETUP_COMPLETED, true));
         connectedPlayers.switchActive();
     }
 }
