+++ server/src/main/java/com/battleships/Commands/CommandsImpl/StopPlayingCommand.java
@@ -3,14 +3,13 @@ package com.battleships.Commands.CommandsImpl;
 import com.battleships.commands.AbstractCommand;
 import com.battleships.player.ConnectedPlayers;
 
-public class EmptyCommand extends AbstractCommand {
-
-    public EmptyCommand(Object value) {
+public class StopPlayingCommand extends AbstractCommand {
+    public StopPlayingCommand(Object value) {
         super(value);
     }
 
     @Override
     public void execute(ConnectedPlayers connectedPlayers) {
-        System.out.println("Unknown command was send to server.");
+
     }
 }
